requirements.txt
anthropic>=0.40.0
aiohttp>=3.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
slack-sdk>=3.27.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.0
